chore(tickables): make tickable module
