+++ tests/test_tickables.py
@@ -1,0 +1,6 @@
+from froggytv.tickables import Noop
+
+
+class TestNoop:
+    def test_noop_does_nothing(self):
+        assert Noop().tick(0) is None
